third_party/WebKit/LayoutTests/editing/selection/caret-mode-paragraph-keys-navigation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="target">This tests that arrow keys navigate through a paragraph as expected when in caret browsing mode, also with shift and ctrl modifiers.</p>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that keys navigation through a paragraph work as expected when in caret mode, also with shift and ctrl modifiers.");

if (window.testRunner && window.internals) {
  testRunner.overridePreference("WebKitEnableCaretBrowsing", true);
  testRunner.dumpAsText();
  internals.settings.setEditingBehavior('mac');

  var target = document.getElementById("target");
  target.focus();
  window.getSelection().collapse(target.childNodes.item(0), 15);

  // check left/right arrows
  shouldBe("getSelection().baseOffset", "15");
  eventSender.keyDown("ArrowRight");
  shouldBe("getSelection().baseOffset", "16");
  eventSender.keyDown("ArrowLeft");
  shouldBe("getSelection().baseOffset", "15");

  // check left/right arrows + Shift
  eventSender.keyDown("ArrowRight", ["shiftKey"]);
  eventSender.keyDown("ArrowRight", ["shiftKey"]);
  eventSender.keyDown("ArrowRight", ["shiftKey"]);
  shouldBe("getSelection().baseOffset", "15");
  shouldBe("getSelection().extentOffset", "18");
  eventSender.keyDown("ArrowLeft", ["shiftKey"]);
  shouldBe("getSelection().baseOffset", "15");
  shouldBe("getSelection().extentOffset", "17");

  // check left/right arrows + Ctrl/Option
  var modifierForWordGranularity = navigator.platform == "MacIntel" ? "altKey" : "ctrlKey";
  eventSender.keyDown("ArrowRight", [modifierForWordGranularity]);
  shouldBe("getSelection().baseOffset", "21");
  eventSender.keyDown("ArrowLeft", [modifierForWordGranularity]);
  shouldBe("getSelection().baseOffset", "16");

  // check left/right arrows + Ctrl/Option + Shift
  eventSender.keyDown("ArrowRight", [modifierForWordGranularity, "shiftKey"]);
  eventSender.keyDown("ArrowRight", [modifierForWordGranularity, "shiftKey"]);
  shouldBe("getSelection().baseOffset", "16");
  shouldBe("getSelection().extentOffset", "26");
  eventSender.keyDown("ArrowLeft", [modifierForWordGranularity]);
  shouldBe("getSelection().baseOffset", "22");
}

</script>
</body>
</html>
